Make the default for 'stop' and 'restart' to wait
The recommended behavior for stopping or restarting the daemon
should be to wait for it to be completed. The old default was not
to wait and in order to wait the '--wait' flag had to be set.

Here, we invert the logic and wait by default and if the user does
not want to wait , they can set the '--no-wait' flag. This is the
safer default behavior