Temporarily restrict the version of tornado to 4.5.3
On March 5 2018, tornado v5.0 was released, which would get installed
automatically, however, this breaks another dependency, circus.
As of this writing circus has not yet fixed this in a release or
locked down the tornado dependency and so we have to lock it.
Without this circus will not be able to start the daemon, which is
rather crucial