Ensure that check_outputs in on_finish of process does not raise
Previously, in the on_finish call of a Process, the outputs would
be validated by calling _check_outputs, which would raise if the
validation failed. This was called regardless of the result that
was passed as the return value. This meant that even when a process
returned a non-zero value, indicating it should got to the FINISHED
state with a non-zero FINISH_STATUS, indicating failure, the outputs
would still be checked even though it would be to be expected that
they would not be correct.

The solution for this happened in plum so this commit merely updates
the requirement and function signature. The on_finish method now
takes an additional argument 'successful' which if false, the outputs
will not be checked. On the other hand, if successful is True, the
outputs will be checked, however, if they are not valid an exception
will not be raised but the transition to FINISHED state is rejected
and another transition to FINISHED is scheduled but this time with
the 'successful' argument set to False