Make verdi daemon stop database backend independent
When the database needs to be migrated, the user will be presented with
the instruction to call verdi daemon stop to make sure it is not running.
To not have that raise the same error, it needs to run without loading
the database environment. It only needs the configuration, which should
also not need the backend. Note that we also had to move the definition
of the USE_TZ because importing it led to the check of AIIDADB_PROFILE
being set which requires the backend to be loaded, however, USE_TZ does
not depend on the profile nor backend. So we just moved it out