Implement the concept of hidden verdi sub commands
Certain sub commands of a VerdiCommandWithSubcommand may need to
be hidden as in the case of 'verdi daemon _start_circus'. To
facilitate this introduce the property 'valid_subcommands' that
will return the valid_subcommands dictionary excluding those keys
that start with an underscore. This means that a command with a
leading underscore key, will be considered hidden and will not be
autocompleted, but it will still be callable from the command line