Change verdi profile list to use colors by default
The method of 'verdi profile list' to indicate the default profile
was not very clear, because it used symbols in the itemization for
all profiles, just the default profile had a different symbol, and
colors were switched off by default. The '--colors' switch could be
used to turn it on.

I have reversed that logic and made colors the default, providing
the switch '--no-colors' to switch it off and I have removed all
itemization symbols, except for the default profile. The output
now matches exactly that of for example 'git branch' which is a
lot clearer