Revert the DEFAULT_TRANSPORT_INTERVAL to 30
This was temporarily set to 5. because for a larger interval the daemon
tests on Travis would fail. The final JobCalculation would not properly
be put into the ProcessState.FINISHED causing the loop that checks that
all calculations are terminated, to time out. With the new circus daemon
and fixes to the ProcessState system, this problem now seems to have
been solved