Add logging handler for pika with loglevel set to ERROR
Currently the pika adapter fires a WARNING log message everytime
the top level submit function is called to submit a process to the
daemon for launching. This is because after submission the connection
to rmq is closed but the event loop is kept running. This is intentional
as the event loop is not only running the rmq connection but also the
tasks of the daemon runner itself. For now this a quick and dirty
solution to suppress these warning log messages