Add support for site_tolerance in CifData._get_aiida_structure
The CifParser of pymatgen allows to set the parameter site_tolerance
which will be used to determine whether two sites overlap and if
created by symmetry they should be merged as one. By default this
value is 1E-4 in v4.5.3, which is rather tight and often leads to
a created structure with overlapping atoms. To prevent this one
should be able to specify this parameter, however, unfortunately
it is not supported by the other library 'ase' that AiiDA supports
internally for structure generation from CifData nodes.