Refactor circus functionality into CircusClient
This cleans up the code in verdi daemon commands by centralizing all
the code that interacts with the circus deamon through a CircusClient
instance that can be configured for a particular profile. I had to
get rid of the `@only_if_daemon_pid` decorator on the function that
called the circus endpoint for any given daemon, because this prevented
the implementation of `--all` flags such as for `verdi daemon status`.
The decorator would call `sys.exit(1)` straight away if a daemon was
not running. Instead we check manually if the daemon is running and
just return a status if it is not instead of killing the process