Ensure the correct profile is used when using -p for verdi daemon start
The `verdi daemon start` command delegates the actual starting of the
daemon to the private `_start_circus` command. However, since that is
its own proper verdi command, it will take the active profile from either
the default profile or the one passed with the `-p` option. This means that
when `verdi daemon start` is called with the `-p` option, it has to pass it
through to the `_start_circus` command call.