Fix prospector warnings
The `pycifrw_from_cif` function in `aiida.orm.data.cif` was triggering
a `too complex` warning by the `mccabe` tool, which is a tool that
determines the complexity of a function and is run by prospector.
Safe for raising the allowed complexity error, I did not find a way
to disable this analysis for that only function like you can do for
pylint warnings. Instead I had to disable it completely in the
prospector config