Detect occupation violation in pymatgen cif to structure parsing
The CifParser of pymatgen will check the atomic occupations of the
parsed cif and if they exceed unity, a warning will be emitted and
the parsing of the structure fails. It is possible to give a certain
tolerance to allow invalid occupations through the occupation_tolerance
of the CifParser constructor. Any occupations between one and this
tolerance will be rounded to one.

Pymatgen does not distinguish between various parsing errors, any
failure just raises a ValueError. However, we would like to distinguish
between well defined errors such as the incorrect occupations case.
As a solution, if the parsing fails, we attempt a second time, this time
setting the occupation tolerance to an unrealistic high value. If the
parsing now succeeds, the original failure was due to the occupations
being exceeded and we raise a specific error InvalidOccupationsError.
If the second parse attempt also failed, there was some unknown parsing
error and we simply raise a ValueError.