Remove the None default from the cif to structure converters
These converter functions are optional inlines, which means that
if they are called with `store=True`, AiiDA will try to store the
input arguments, but if the `parameters` were not specified, it
will try to store the `None` default, which of course fails.
The solution is to use `**kwargs`. With this a user is not required
to specify an empty dictionary or `ParameterData` node, and the
function can be used both as inline and as regular function.