Obtain circus controller endpoint ports from OS instead of profile
The previous solution was generating a hardcoded circus port for each
profile which was then being used for the three endpoints of the circus
daemon. However, since these profile ports were only guaranteed to be
unique between the profiles of an aiida installation, i.e. one config.json
file, and not system wide, there was a large chance for ports overlapping
and the daemon not being able to start. Not to mention any other processes
on the system that might happen to run on the same ports.

Instead we now determine the port by requesting an available one from the
operating system. The only problem is that once the circus daemon is running
with a given port for its controller, the daemon client needs to remember
this value to be able to communicate for as long as the daemon is running.
We currently write it to a file tied to the profile, however, this means
it has to be destroyed when the daemon stops, just as the pid file. However,
the DaemonClient has no control over the daemon itself and circus only takes
care of deleting the pid file. It is thus possible that if the daemon is not
shutdown nominally, the port file remains and the next time the daemon is
started the port value that it contains may have gone stale and is now being
used by a different process.