Make 'state' and 'job_state' compound projection for 'calculation list'
The default set of projections for `verdi calculation list` should not be
wider than 80 columns roughly, otherwise the lines for people with default
shell widths will just wrap. However, with the recent changes of homogenizing
the spheres of WorkCalculations and JobCalculations through the process
layer, we want to give `verdi work list` and `verdi calculation list` a
similiar default output. In addition to the default projections of the
former, the calculations also have the calculation and scheduler state which
should remain there by default, but this would make the output to wide. The
solution is to define compound projections:

	* state
	* job_state

The first applies to both `work` and `calculation` and will show the tuple
of `process_state` and `finish_status`. The second only applies to the
`calculation list` and is the tuple of the old calculation state and the
scheduler state.