Tie the lifetime of circus port file to the pid file
To determine whether an existing circus port file is stale, when
requesting the circus port from the DaemonClient, we simply verify
if the daemon is running by checking the existence of the pid file.
If the daemon is not running, an available port is requested from
the operating system, which is written to the port file. This port
file will from then on be used by the daemon as long as it remains
alive