Implement the IPC protocol for the circus daemon
The circus library provides no ways of security when using the TCP
protocol, so that is a vulnerable choice. It also provides support
for the IPC protocol and have the endpoints communicate over sockets.
We implement this protocol for the DaemonClient and set that protocol
as the default. In the future, the desired protocol can be made
configurable on a per installation or even profile basis