New and clean plugin loader implementation
This implementation considers the entry point system to be a first
class citizen and only if that fails reverts to the old internal
system. This allows for a much cleaner definition of the interface
as well as the determination of the `plugin_type_string` attribute
of the `AbstractNode` class that is generated in the metaclass
construction.

The idea is that to construct the correct plugin type string for a
given class, is to first try to reverse map the class path to an
entry point group and name. If that succeeds, it means the class is
external and we get the corresponding internal orm class from the
orm class mapper. If it fails, we fall back on the assumption that
it is an aiida internal class and we assume it is already the correct
orm class path. Finally the expected potential prefixes `aiida.orm.`
and the various `implementation.*` prefixes are stripped.