Add column `process_type` to the DbNode model
With the arrival of the `Process` concept, the class that powers the
new workflow engine, an asymmetry was introduced in relation to how
calculations used to be ran and how they are ran now. Originally, a
JobCalculation was a class that both served as a container to record
the results of the executed calculation. In that sense it truly is an
ORM class as it maps the contents of the database to an object.
However, it also contained all the logic and rules of how to run the
calculation. Now, those two demains have been separated. The ORM class
should only serve as a record of the execution of a calculation, whereas
the Process subclass implements the rules for execution.

The WorkChain, a sublass of Process, implements this new concept and uses
the WorkCalculation ORM class as its database record. However, there is
no way yet to store the identity of the process in the record. The new
process_type column will facilitate this requirement. For the format, we
choose to leverage the entry point system. The entry point manager will
be asked to give the entry point group and name that corresponds to the
Process class and the concatenation will be stored in the process_type.
This is useful because this process_type string uniquely defines the
plugin class and through the plugin loader and factories the class can
be easily reloaded.