Replace the old load_plugin with one that only considers entry points
The load_plugin function attempted first to call the old plugin loader
that was based on the old internal plugin structure and tried to
directly load modules from their modulepath. If that failed there was
a fall back on to the new loader, that leveraged the entry point system
which allowed the loading of classes external to aiida core.

Here, we completely rip out the old method and rewrite the plugin loader
to only allow the loading of plugins through an entry point. To this end
we also added the few remaining internal base classes to the entry point
definition in the setup.py that weren't yet included.

For now, I had to keep an explicit check for `data.Data` in the load_plugin
function and not add that base class to the entry points, because if I
do add it to the entry points, some query tests regarding subclassing
break. This will have to be figured out in the future.