Set the _calc_class attribute for the ContinueJobCalculation
The ContinueJobCalculation inherits from the JobProcess class to
allow continuing a created JobCalculation, however, for this the
_calc_class needs to be set, which is done in the `build` class
method of `JobProcess`. This means that if ContinueJobCalculation
is instantiated directly the _calc_class was not being set. We
override the `_setup_db_record` and set the `_calc_class` based
on the class of the `_calc` input node, which is the calculation
class wrapped by `ContinueJobCalculation`.