Adapt the function signature of BaseFactory and load_plugin
These two functions were designed at the time when plugins were internal
to the aiida core source tree and were loaded as internal modules. To
support external plugins, the plugin system was designed to allow new
plugins to be added through entry points. For a long time the loading
of plugins was a mixture of those two paradigms. Now that we no longer
support loading of plugins through internal modules and everything has
to go through entry points, the old signature of BaseFactory and the
load_plugin function also no longer made sense. The BaseFactory now
just takes an entry point group and name, while the load_plugin function
still operates on a version of the plugin type string of the Node class
and its subclasses.