Increase the default timeout value for circus controller call
The timeout was hardcoded to 0.5 seconds for calls to the circus
daemon controller endpoint, which, when multiple workers were running
could very easily lead to a timeout. The subsequent error was also
not recognized properly by the caller in get_daemon_status, resulting
in an exception. The error is now properly defined and checked for
with a useful error message that is printed. The default timeout
has been set to 3 seconds