Change the SqlAlchemy migration of code hidden attribute to use raw SQL
The original migration was using the SqlAlchemy models to perform the
migration of the 'hidden' attribute of Code nodes to the extras field.
However, this is problematic, because that will break if the model is
ever going to change in the future. Instead we should use raw SQL
statements that are independent on the current model and will always
work.