Use class path as fall back for process_type if no entry point found
Some processes will be run that are not registered as a plugin with
an entry point, so the process_type cannot be set to the entry point
type string. In that case it is better to fall back on the absolute
class path as opposed to nothing at all.

Also made sure to add the process_type to the DummyModel instance
creation such that entries retrieved with the QueryBuilder for the
Django backend will have the process_type attribute.

Added property shortcuts to the node implementations to get the
process_type directly from the node instance