Implement the Node.load_process_class method
With the addition of the process_type field for the Node model, instances
that were ran by a Process and therefore have a process_type value, can now
easily reload the original process class from the node instance by calling
the `load_process_class` method. If there is not process_type value, None
will be returned. If it seems to be a valid entry point string, the class
will be attempted to be loaded through the entry point. Otherwise we try
to import it through importlib.