Ensure that raw CIF content from IcsdDbEntry is UTF-8 encoded
The PHP database backend of ICSD returns the raw CIF data encoded in
the IS-8859-1 format. However, most parsing libraries, such as PyCifRW
that are subsequently used to parse the raw CIF content expect the
format to be UTF-8. If the CIF contains characters that when encoded in
ISO-8859-1 are not recognized by the UTF-8 format, attempting to decode
it as such will raise the UnicodeDecodeError exception. To solve this
we simply decode the raw content as soon as it is retrieved from the URI
and immediately encode it in UTF-8.