Remove `verdi work tree`
This command has been superseded by the `verdi work status` command,
which will print the status of the workchain and that of the calculations
it called. The output will be formatted in a tree like structure.
For an overview of the inputs and outputs of the workchain, a user
can better use the `verdi calculation show` command.