Ensure that Process.kill() cascades down to called children
If kill is called on a Process, it needs to make sure that it also
issues the kill command for all the children processes that it called.
This is achieved by first getting all its called children by following
CALL links out of the calculation node, and then calling kill_process
on the rmq control panel of the process' runner with the pids of the
children calculations

This change also allowed to remove the skips from the unittest that
test the cascading of the kill command for processes with children.