Shutdown the DaemonRunner neatly when killed
The killing of the DaemonRunner, which is launched by the start_daemon
command, can be accomplished by two signals, SIGINT and SIGTERM. The
SIGINT will be sent when a keyboard interrupt is fired, for example
when the daemon is being ran in the foreground and the user fires CTL+C.
The SIGTERM is the signal sent by the circus arbiter when it is asked
to quit, which is done by calling `verdi daemon stop`. To make sure that
for both these cases the DaemonRunner shuts down neatly, we simply bind
both signals to a function that when triggered will close the connection
of the DaemonRunner and stop the loop. The loop is a Tornodo library
event loop and should take care of shutting down neatly. Any running
tasks that were picked up from RabbitMQ should be rescheduled by RMQ
when it realizes the subscriber died. When it is revived, i.e. when
the daemon is restarted, these tasks will be redelivered.