Update `verdi calculation kill` to use the ProcessControlPanel
Also ripped out the old JobCalculation.kill method as that should
not be called anymore. Currently for the implementation of the
`verdi calculation kill` method, we use a non blocking ControlPanel
which means it will return futures. This is to ensure that if a
user requests multiple calculations to be killed, they are all fired
at once, because if a remote transport is required, that will only
be made available in certain intervals. Therefore we cannot fire
the kill request one by one, because that would take forever. Instead
we now wait on the future, which should all be actioned in a single
transport run. In the future we can add an option to also not wait
for the confirmation of the futures and immediately return from the
command after the kill requests have been sent.