Reduce the number of state reentering in JobProcess.scheduler_update
Whenever a JobProcess would be in the WAITING state while waiting for
the calculation to finish, the scheduler_update would be called constantly,
inducing a state change from WAITING to WAITING everytime, which would
call the persistence. For environments with a transport queue interval that
is very small, such as for local transports, this would cause a lot of
database hits. By putting a while loop within the WAITING state we prevent
the excessive persisting through the unnecessary state changes