Do not persist Process in on_entering if state is CREATED
With the mechanism of launching a Process that relies on one worker
instantiating the Process, creating the database record, persisting
it before creaating a ContinuationTask, which will be picked up by
another worker (if there is more than one worker), since the persisting
is called with a delay, the two workers may call the persistence at
a similar point in time. This race condition will then lead to a
database IntegrityError for one of the two, which will put the Process
in the excepted state, causing the other one to also fail.
By putting in the check for the CREATED state and omitting the persist
call in that case we avoid this problem.