Implement ExecuteProcessAction to not send launch over RabbitMQ
The ExecuteProcessAction defined in plumpy will first send a launch
task over RMQ to create the process and with a callback when its
done, submit a ContinuationTask and return its own future. This
mechanism, however, if there is more than one worker subscribed to
launch and continuation tasks, would end up in two workers having
the same process in memory and picking up messages for it. This
would lead to race conditions and database integrity errors.

We reimplement the ExecuteProcessAction in aiida-core and instead
of sending the launch task over RabbitMQ, we create the Process
instance locally, persist it amd then close it, which means to
remove all the rpc subscribers it had registered. Then we simply
send the ContinuationTask over RMQ and return the pk as the result.
This guarantess that there are not more than workers picking up
messages for the same process