Serialize the context of a WorkChain before persisting
The user is free to populate the context with Node instances, which
means that if the WorkChain needs to be persisted that the nodes
need to be serialized. Since the nodes are not necessarily stored
upon calling save_instance_state, we also store them if they were
not yet stored. We also use this opportunity to replace the ad hoc
serialization in the pickle persister and use the new more complete
serializer and deserializer.