Make the node count optional for verdi group list
For big databases the verdi group list command gets very slow when
the groups contain a lot of nodes. Often one is just interested in
the names and pks of the groups and does not want to have to wait
multiple seconds to also get the node count. Instead there now is
a flag -C to indicate to also print the node count