Various fixes to prevent RMQ channel and connection from closing
All subscribers to a specific queue of RabbitMQ get a heartbeat with
a default timeout of 60 seconds. If the heartbeat callback is not actioned
before the timeout expires, RabbitMQ is going to assume that the subscriber
died and is going to close its connection. Combined with the unlimited task
prefetch count that we set for the DaemonRunner, it could take all the tasks
it wanted and it could easily lead to the situation where it would be to busy
to action the heartbeat callback and the connection would be closed. After
which all the tasks that the runner had, would fall over as soon as they tried
to communicate with RabbitMQ. To deal with this situation, a few things have
been changed in kiwipy, plumpy and aiida-core.

 * The default queue heartbeat timeout has been set to 600 seconds
 * The task prefetch count has been limited to 20
 * The auto_delete from the messages and tasks queues have been removed
 * The expiry time of the task queue has been set to one week
 * The expiry time of the message queue has been set to one minute

These measures should limit the chance that the heartbeat timeout is reached,
causing the connection to be closed. Additionally, changes have been added to
cancel the current tasks of a runner, when the connection does get closed.
The change of the auto_delete and the increased expiry times of the queues
should prevent that existing tasks get deleted too soon by RabbitMQ when the
daemon is shutdown.