Changed SQLA Node to get live data from db
There was an issue where because a node was in a Session, if any of the attributes
of a Node were changed by an external process the getters would not see the change.

I've put in an session.expire from SQLA to handle these cases.  Also refactored
AbstractNode a little.