Reraise exceptions caught in submit_calc to SubmitJob TransportTask
The old execmanager implementation of `submit_calc` would catch any
exceptions during the submit process, set the calculation state to
SUBMISSIONFAILED and return True or False. Since the exception was
not reraised, the TransportException of the SubmitJob TransportTask
of the JobProcess was not being triggered, which caused the JobProcess
to go to the next step of UpdateSchedulerState transport task which
would except because it assumes that the job was successfully
submitted. Reraising the exception in submit_calc fixes the problem.