Ensure that verdi work list also lists FunctionCalculations
Workfunctions should also be considered by verdi work list and so their
database records, the FunctionCalculation class, should also be queried
just like the WorkCalculation. Unfortunately, the QueryBuilder does not
yet support querying for a tuple of classes in its append method. Two
separate queries is also no option, since the ordering then becomes
problematic. A temporary solution is to query for the Calculation base
class and add filters on the `type` column too match that of the
FunctionCalculation and WorkCalculation