Enable passing of JobCalculation class to launch functions
This prevents the user from having to call JobCalculation.process() manually
before passing it to the launch function and makes it homogenous with passing
a WorkChain or a workfunction. In adding this functionality also cleaned all
the unnecessary clutter with top level functions dealing with instantiating
the process. Have reduced it to a single function that does all the necessary
checks.