Fix the problem with socket filename size limitation for circus IPC protocol
Currently, we are using the IPC protocol as the default for the daemon to
communicate with the endpoints of the circus controller. The sockets used
to be stored in the AiiDA config folder, however, UNIX has a built in byte
limit of 107 on socket filenames and given that the config folder can be
nested arbitrarily deeply, this limit was easily exceeded. The solution is
to store them in a temporary directory. This solution should always avoid the
limit comfortably but adds slightly more complexity. A temporary folder is
generated by the OS and is only accesible by the user who created it, so
the security is guaranteed here, which is what the IPC protocol was supposed
to solve in the first place over the TCP protocol.