Changed AiiDA user to use new backend concept
The old system (which is still in place for most AiiDA types) uses
an abstract that is subclassed by the concrete backendclass e.g.
AbstractUser and DjangoUser and then DjangoUser is then imported
at load_dbenv.  This has a lot of problems, including that you have
to load_dbenv at the right time, it prohibits having more than one
backend in play at a time and it means that we use global variables
which make debuggin harder.

This bring User into conformance with the new system which is so
far used just by Log.