Documentation: advanced concepts of workchain design
Wrote extensive documentation on all the features of the WorkChain
and how to write robust, maintainable and modular workchains. Described
all the features of the ProcessSpec with the newly added Ports and
PortNamespaces, the context and how to add futures, the reporting system
and how to abort workchains and set their finish status.