Add the pyproject.toml file which replaces `setup_requires`
Starting `v10.0.0`, pip supports PEP 518 through the `pyproject.toml`
file, which can be used to define dependencies that are required by the
build process. We chose to use this mechanism over the `setup_requires`
keyword in the `setup` call of `setup.py`, because that causes problems
on older standard python version on Mac OS X as that goes through setuptools
instead of pip.

To ensure that the version of `reentry` specified in the `pyproject.toml`
and the `setup_requirements.py`, I also add a pre-commit test.