Fix verdi export create to be able to handle multiple values for an option
Click does support multiple values for an option through the `multiple` keyword
of the option constructor, however, this requires the user to repeat the flag
for each value, i.e.:

    --nodes 1 --nodes 2 --nodes 15

For big lists, this is impractical and untenable. Therefore we introduce a custom
MultiValueOption that instead supports the following notation:

    --nodes 1 2 15

Since these options are greedy, they can clash with cli commands that also take
arguments. If the greedy option is used, the argument will be mistaken for another
option value. In this case, the MultiValueOption respects the endopts marker '--'
that is the standard for POSIX command line interfaces. Anything after this marker
will be considered as arguments by the parser.