Fixed class_loader (LOUDER) issue #1435
Made a couple of changes in plumpy to make this more obvious and now the
class loader (renamed to object loader because it can also load functions,
constants, etc) is a subclass of the one in plumpy.  Its class name is
being saved in the bundle and it is being correctly loaded up when loading
a saved process.