Implement the `submit_test` functionality for ProcessBuilder
The `JobProcess` overrides the `get_builder` method to return a subclass
of the `ProcessBuilder`, the `JobProcessBuilder` which has an additional
method called `submit_test`. This method will call through to the same
method of the `JobCalculation` class, which will create a folder in the
current working directory with the input files and scripts as they would
have been created on the remote computer, were the builder have been
submitted for real. This gives the user a chance to inspect the generated
files without having to store the calculation or inputs nodes and actually
submitting the job to the remote.