Add integration workflow: the most contrive Reverse Polish Notation solver ever
This implements a command line interface script that can generate or take
a random expression in Reverse Polish Notation and evaluate it. However,
it will evaluate it in two ways.

 * The normal way, with a straightforward python implementation
 * The hard way, through a dynamically generated nested workchain

The expressions are limited to use integers and only summation, multiplication
and exponentiation. Exponentiation is modeled through launching sub workchains,
summation is done through JobCalculations or workfunctions, conditional on the
sign of the operand and multiplication is modeled by a while loop. In this
manner all major constructs of a workchain are used and in a random configuration.
By generating random expressions, running them with both evaluators and comparing
the result, we can check the workflow engine.