Update the setup_requirements.py to most recent stable versions
In order to guarantee the highest chance of a build that won't break
by installing from PyPi is to fix the version numbers in the setup
requirements, such that if dependencies break our code, the released
versions on PyPi will still work. Therefore we update here all
dependencies to the most recent version that do not break our code.
Current exceptions that have to be pinned at an older version are:

 * django
 * django-extenstions
 * sqlalchemy
 * tornado
 * ase
 * pycifrw
 * ipython

All other packages have been updated to the most recent version.
As a consequence of this decision, we can also remove the now
obsolete requirements.txt