Add support to run the RPN tests over the daemon
This requires adding a timeout mechanism to check after submission
when the workchain is terminated, and timing out if it takes too
long. Note that this can actually happen if an expression is generated
that will end in more processes in parallel than the single worker
of the daemon has, resulting in a lockdown. This issue is currently
being worked on but might still occur at the moment of writing.