Change the test_rpn.sh script to use predefined expressions
For these expressions we know that the timeout will not be
hit. We lose the randomness of testing but the tests were
timing out too often and would otherwise take too long.
The script is written such that just commenting out the line
with the definition of EXPRESSIONS variable, will return to
the old behavior of running with randomly generated expressions.