Make sure that kwargs do not override existing ports in FunctionProcess
The FunctionProcess inherits the ProcessSpec from Process which already
defines some input ports such as `label` and `description`. In the build
method of the FunctionProcess it creates additional ports dynamically
based on the passed args and kwargs of the function. These were overriding
any existing ports with default parameters. For the case of the `label`
and `description` ports, overriding with the defaults, meant losing the
correct `valid_type` and `non_db` attribute values. This in turn caused
setting the `label` and `description` for a workfunction to except during
the creation of the database record.