Ensure that getters of ProcessBuilder return actual values
The old solution returned an instance of ProcessBuilderDictionaryInput
which means that simple things like value comparisons were broken. The
user would always expect to get the actual value that is stored under
the attribute and not some proxy. To solve this problem while still
maintaining the dynamic creation of doc strings and support for the
nested PortNamespace, we turn the ProcessBuilderInputDict into a
simple ProcessBuilderNamespace that is essentially a mapping. Upon
instantiating it will dynamically create the getter and setter properties
based on the ports of the process spec of the process.