Separate the `build` and `run` steps for a workfunction
When a function is decorated with the workfunction decorator and the
function is subsequently run, first the corresponding FunctionProcess
needs to be built, with the ProcessSpec built to mirror the signature
of the wrapped function and when that is done, construct the process
instance passing the actual inputs. These steps are two separate
concepts but they were both executed only upon execution of the function.
However, the definition of the ProcessSpec can already be completed as
soon as the workfunction decorator is applied. Any inconsistencies, e.g.
the definition of variadic arguments which are not supported, can then
already be caught at the build step.