transport.ssh: execute using login shell
use login shell to execute all remote SSH commands

This has the advantage that all commands will be executed in the same
environment as seen by a user loggin in onto the machine
(e.g. on most UNIX systems, .profile will be sourced for login shells,
while it won't be sourced for non-login shells).

Additionally, calling `bash -l -c '<cmd>'` instead of `<cmd>`
means that the bash shell will be used, independent of the default
shell on the system.
This is advantageous on systems, where bash is not the default shell
(since our remote commands work only in bash).