Added back the last daemon status update
This was no longer changing in verdi calculation list because the old
daemon has been replaced by an event based daemon.

To give some of this functionality back we now set the global settings
timestamp every time a job calculation saves a checkpoint.  Note, there
is a consequences of this that may appear to be surprising:

The 'daemon' timestamp will be updated even if the JobProcess is ran
in the local interpreter.

I feel this is acceptable but if we feel it is misleading perhaps we can
change the string printed to better reflect what this timestamp means.