Fix bug that circumvented check whether the daemon is running
The bug was in the decorator `only_if_daemon_pid` because the property
`get_daemon_pid` was changed to a method in a previous commit. As a result
the conditional was never hit and some commands that depended on the daemon
running such as `incr`, `decr` and `restart` were executing even if the
daemon was not running.