Remove the old daemon task timestamp getter and setter functions
The new daemon no longer operates based on tasks and so the timestamp
for the last iteration of those tasks have become obsolete. The only
task that still remains is the legacy workflow ticker. Therefore we
remove the generic daemon timestamp setters and replace them with
a version that is explicit to set and get the timestamp for the
legacy workflow ticker. This is still needed to for the mechanism that
ensures that the same workflows are not ticked multiple times at once.