Add trigger to Process.on_entered to write state change timestamp
A call is added to the Process.on_entered hook, which will update
the value of a setting in the DbSetting table with the current
timestamp. The exact key will depend on the type of the calculation
node used by the process, such that a distinction can be made between
'normal' calculations like a job calculation and 'work' calculations
such as WorkCalculation nodes. These timestamps can then be used in
for example `verdi calculation list` and `verdi work list` to notify
the user when the last time was that a process changed its state.
This is useful to give a sense whether something is actually happening.