Print statement with last process state change in verdi list commands
This concerns `verdi calculation list` and `verdi work list`. We add
the call to a function that will print information about the last time
that a process with the corresponding type changed its state, to give
the user an indication whether things are actually progressing. It
will also print a warning if the daemon is not running.